inbound/quic: fix an NAT issue caused by missing source address and stream ID

+++ leaf/src/session.rs
@@ -25,7 +25,7 @@ impl std::fmt::Display for Network {
     }
 }
 
-pub type StreamId = u16;
+pub type StreamId = u64;
 
 #[derive(PartialEq, Eq, Hash, Clone, Copy, Debug)]
 pub struct DatagramSource {
